apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-multipart==0.0.20
python-dotenv==1.0.1
httpx==0.28.1
anthropic==0.43.0
chromadb==0.6.3
sentence-transformers==3.3.1
PyMuPDF==1.25.3
reportlab==4.2.5
tqdm==4.67.1
pdfplumber==0.11.4
pillow==11.1.0
pytest==8.3.4
pytest-asyncio==0.25.0
